+++ src/Haskellorls/Config.hs
@@ -74,7 +74,6 @@ data Config = Config
     tabSize :: Int,
     tree :: Bool,
     width :: Int,
-    noQuote :: Bool,
     zero :: Bool,
     dired :: Bool,
     currentWorkingDirectory :: RawFilePath,
@@ -195,7 +194,6 @@ mkConfig env Option {..} = Config {..}
       Format.SINGLECOLUMN -> 1
       Format.LONG -> 1
       _ -> fromMaybe 80 $ oWidth <|> Env.columnSize env
-    noQuote = False
     zero = oZero
     dired = oDired && format == LONG && not tree
     toTTY = Env.toTerminal env