+++ src/Haskellorls/Formatter/Name.hs
@@ -1,7 +1,5 @@
 module Haskellorls.Formatter.Name
   ( colorizedNodeName,
-    colorizedNodeNameWrapper,
-    nodeNameWrapper,
     nodeName,
   )
 where
@@ -12,29 +10,19 @@ import qualified Data.Text.Encoding as T
 import Haskellorls.Class
 import qualified Haskellorls.Config as Config
 import qualified Haskellorls.Formatter.Attribute as Attr
-import qualified Haskellorls.Formatter.Escape as Escape
-import qualified Haskellorls.Formatter.Quote as Quote
 import qualified Haskellorls.Formatter.WrappedText as WT
 import qualified Haskellorls.LsColor as Color
 import qualified Haskellorls.NodeInfo as Node
 import System.FilePath.Posix.ByteString
 
--- FIXME: Quote with double quotes even if doesn't contain any double quote.
-colorizedNodeNameWrapper :: Config.Config -> Color.LsColors -> Node.NodeInfo -> [Attr.Attribute WT.WrappedText]
-colorizedNodeNameWrapper config lc nd = Quote.quote (Quote.quoteStyle config) $ colorizedNodeName config lc nd
-
 colorizedNodeName :: Config.Config -> Color.LsColors -> Node.NodeInfo -> Attr.Attribute WT.WrappedText
 colorizedNodeName config c@(Color.Options {..}) nd = Attr.Name $ WT.WrappedText (left' <> l <> right' <> wtPrefix) wtWord (wtSuffix <> left' <> r <> right')
   where
-    WT.WrappedText {..} = Attr.unwrap $ WT.modify (Escape.escape config) <$> nodeName config nd
+    WT.WrappedText {..} = Attr.unwrap $ nodeName config nd
     left' = Color.unSequence $ fromMaybe "" left
     right' = Color.unSequence $ fromMaybe "" right
     l = maybe "" Color.unSequence $ nd `Color.lookup` c
     r = Color.unSequence $ fromMaybe "" reset
-
--- FIXME: Quote with double quotes even if doesn't contain any double quote.
-nodeNameWrapper :: Config.Config -> Node.NodeInfo -> [Attr.Attribute WT.WrappedText]
-nodeNameWrapper config node = Quote.quote (Quote.quoteStyle config) $ WT.modify (Escape.escape config) <$> nodeName config node
 
 nodeName :: Config.Config -> Node.NodeInfo -> Attr.Attribute WT.WrappedText
 nodeName config@(Config.Config {hyperlink, hostname}) node =